pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "installer_builder2"
description = "Build distributable applications"
version = "0.1.3"
authors = [
    { name = "Christopher Toth", email = "q@q-continuum.net" }
]
dependencies = [
    "attrs",
    'innosetup_builder',
    'nuitka',
    'ordered_set', # For faster Nuitka compiles
]